More work on collisions. Worked on a bug where the player could get stuck if they jumped into corners - ended up making it worse for the time being, but closer to fixed in the long run.
Committing with a known bug: jumping into corners allows you to clip into the walls. This is not a feature. Fairly sure this is due to the fact that collisions are checked twice: once for vertical, once for horizontal. This doesn't catch diagonal collisions where the player is alright on each axis, but not on the vector.